fix issue with np int and float not existing

+++ src/arcgis/features/geo/_io/fileops.py
@@ -1187,11 +1187,9 @@ def _pyshp_to_shapefile(df, out_path, out_name):
                             shpfile.field(name=c, size=255)
                         elif isinstance(df[c].loc[idx], (int)):
                             shpfile.field(name=c, fieldType="N", size=5)
-                        elif isinstance(df[c].loc[idx], (np.int, np.int32)):
+                        elif isinstance(df[c].loc[idx], (int, np.int32)):
                             shpfile.field(name=c, fieldType="N", size=10)
-                        elif isinstance(
-                            df[c].loc[idx], (np.float, np.float64, np.int64)
-                        ):
+                        elif isinstance(df[c].loc[idx], (float, np.float64, np.int64)):
                             shpfile.field(name=c, fieldType="F", size=19, decimal=11)
                         elif (
                             isinstance(
